third_party/blink/web_tests/flag-specific/highdpi/fast/css/nested-percent-height-on-replaced-expected.txt

FAIL:
Expected 370 for width, but got 371. 
Expected 170 for height, but got 171. 
Expected 355 for width, but got 356. 
Expected 155 for height, but got 156. 

<div style="overflow: scroll; height: 200px; width: 400px; background-color: salmon">
    <div style="overflow: scroll; height: 100%; background-color: purple" data-expected-width="385" data-expected-height="185">
        <div style="overflow: scroll; height: 100%; background-color: purple" data-expected-width="370" data-expected-height="170">
            <div style="border: 0; height: 100%; width: 100%; background-color: orange" data-expected-width="355" data-expected-height="155"></div>
        </div>
    </div>
</div>
FAIL:
Expected 370 for width, but got 371. 
Expected 170 for height, but got 171. 
Expected 355 for width, but got 356. 
Expected 155 for height, but got 156. 

<div style="overflow: scroll; height: 200px; width: 400px; background-color: salmon">
    <div style="overflow: scroll; height: 100%; background-color: purple" data-expected-width="385" data-expected-height="185">
        <div style="overflow: scroll; height: 100%; background-color: purple;line-height: 0" data-expected-width="370" data-expected-height="170">
            <iframe style="border: 0; height: 100%; width: 100%; background-color: orange" data-expected-width="355" data-expected-height="155"></iframe>
        </div>
    </div>
</div>
